pyproject.toml
[tool.poetry]
name = "meduzzen-internship"
version = "0.1.0"
description = ""
authors = ["Andrew Kabanov <kbnandrii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pytest = "^8.2.2"
httpx = "^0.27.0"
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
databases = "^0.9.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
redis = "^5.0.7"
pytest-asyncio = "^0.23.7"
passlib = "^1.7.4"
aioredis = "^2.0.1"
setuptools = "^70.2.0"
asyncio-redis = "^0.16.0"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.8.0"
python-jose = "^3.3.0"
cryptography = "^42.0.8"
anyio = "^4.4.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pre-commit = "^3.7.1"

[tool.black]
line-length = 110
target-version = ['py36', 'py37', 'py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
